@@ -1,11 +1,15 @@
+import logging
 from typing import Any
 
 import httpx
 
+logger = logging.getLogger(__name__)
+
 
 def oidc_identity_provider_args_from_discovery_url(
     discovery_url: str,
-) -> dict[str, Any]:
+    client_secret: str | None = None,
+) -> dict[str, Any] | None:
     """Build a dictionary of arguments for a Keycloak OIDC Identity Provider.
 
     This helper function is used to simplify the process of registering an OIDC identity
@@ -14,19 +18,25 @@ def oidc_identity_provider_args_from_discovery_url(
     :param discovery_url: The full URL to the OIDC discovery endpoint
         (e.g. https://accounts.google.com/.well-known/openid-configuration)
     :type discovery_url: str
-
-    :raises RuntimeError: If there is an error fetching or parsing the discovery
-        document.
+    :param client_secret: Optional client secret for client_secret_basic authentication.
+        If provided, uses client_secret_basic instead of private_key_jwt.
+    :type client_secret: str | None
 
     :return: A dictionary of arguments that can be passed to the constructor of a
-        keycloak.oidc.IdentityProvider Pulumi resource.
-    :rtype: Dict[str, Any]
+        keycloak.oidc.IdentityProvider Pulumi resource, or None if the metadata URL
+        is inaccessible.
+    :rtype: Dict[str, Any] | None
     """
     try:
         oidc_provider_metadata = httpx.get(discovery_url, timeout=10).json()
     except (httpx.RequestError, ValueError) as e:
-        msg = f"Unable to fetch and parse OIDC discovery document from {discovery_url}"
-        raise RuntimeError(msg) from e
+        logger.warning(
+            "Unable to fetch OIDC discovery document from %s: %s. "
+            "Skipping this provider.",
+            discovery_url,
+            e,
+        )
+        return None
     keycloak_arg_map = {
         "authorization_endpoint": "authorization_url",
         "token_endpoint": "token_url",
@@ -54,12 +64,23 @@ def oidc_identity_provider_args_from_discovery_url(
         oidc_idp_args["default_scopes"] = " ".join(required_scopes)
     else:
         oidc_idp_args["default_scopes"] = " ".join(required_scopes)
-    if (
-        "token_endpoint_auth_methods_supported" not in oidc_provider_metadata
-        or "private_key_jwt"
-        not in oidc_provider_metadata["token_endpoint_auth_methods_supported"]
-    ):
-        msg = f"OIDC provider at {discovery_url} does not support private_key_jwt client auth method"  # noqa: E501
-        raise RuntimeError(msg)
-    oidc_idp_args["extra_config"] = {"clientAuthMethod": "private_key_jwt"}
+    if client_secret:
+        if (
+            "token_endpoint_auth_methods_supported" in oidc_provider_metadata
+            and "client_secret_basic"
+            not in oidc_provider_metadata["token_endpoint_auth_methods_supported"]
+        ):
+            msg = f"OIDC provider at {discovery_url} does not support client_secret_basic client auth method"  # noqa: E501
+            raise RuntimeError(msg)
+        oidc_idp_args["client_secret"] = client_secret
+        oidc_idp_args["extra_config"] = {"clientAuthMethod": "client_secret_basic"}
+    else:
+        if (
+            "token_endpoint_auth_methods_supported" not in oidc_provider_metadata
+            or "private_key_jwt"
+            not in oidc_provider_metadata["token_endpoint_auth_methods_supported"]
+        ):
+            msg = f"OIDC provider at {discovery_url} does not support private_key_jwt client auth method"  # noqa: E501
+            raise RuntimeError(msg)
+        oidc_idp_args["extra_config"] = {"clientAuthMethod": "private_key_jwt"}
     return oidc_idp_args
