@@ -653,9 +653,10 @@ airbyte_helm_release = kubernetes.helm.v3.Release(
                 },
                 "jobs": {
                     "resources": {
-                        "limits": {"memory": "5Gi", "cpu": "1000m"},
-                        "requests": {"memory": "5Gi", "cpu": "1000m"},
-                    }
+                        "limits": {"memory": "15Gi", "cpu": "4000m"},
+                        "requests": {"memory": "10Gi", "cpu": "2000m"},
+                    },
+                    "kube": {"annotations": {"karpenter.sh/do-not-disrupt": "true"}},
                 },
             },
             "postgresql": {"enabled": False},
@@ -676,7 +677,10 @@ airbyte_helm_release = kubernetes.helm.v3.Release(
                 "replicaCount": 2,
                 "deploymentStrategyType": "RollingUpdate",
                 "podLabels": k8s_global_labels,
-                "resources": default_resources_definition,
+                "resources": {
+                    "requests": {"memory": "1Gi", "cpu": "300m"},
+                    "limits": {"memory": "2Gi", "cpu": "500m"},
+                },
                 "httpIdleTimeout": "20m",
                 "extraEnv": [  # How long to attempt new source schema discovery
                     {"name": "READ_TIMEOUT", "value": "30m"},
@@ -686,13 +690,20 @@ airbyte_helm_release = kubernetes.helm.v3.Release(
                 "enabled": True,
                 "hpa": {"enabled": True},
                 "podLabels": k8s_global_labels,
-                "resources": default_resources_definition,
+                "resources": {
+                    "requests": {"memory": "500Mi", "cpu": "100m"},
+                    "limits": {"memory": "1Gi", "cpu": "200m"},
+                },
+                "syncJobMaxTimeoutDays": 15,
             },
             "workloadLauncher": {
                 "enabled": True,
                 "hpa": {"enabled": True},
                 "podLabels": k8s_global_labels,
-                "resources": default_resources_definition,
+                "resources": {
+                    "requests": {"memory": "500Mi", "cpu": "100m"},
+                    "limits": {"memory": "1Gi", "cpu": "200m"},
+                },
             },
             "metrics": {
                 "enabled": False,
