@@ -973,38 +973,42 @@ env_vars = {
     "ALLOWED_HOSTS": '["*"]',
     "AWS_STORAGE_BUCKET_NAME": f"ol-mitlearn-app-storage-{env_name}",
     "CANVAS_PDF_TRANSCRIPTION_MODEL": "gpt-4o",
+    "CONTENT_FILE_EMBEDDING_CHUNK_OVERLAP": 51,
+    "CONTENT_FILE_EMBEDDING_CHUNK_SIZE": 512,
     "CORS_ALLOWED_ORIGIN_REGEXES": "['^.+ocw-next.netlify.app$']",
     "CSAIL_BASE_URL": "https://cap.csail.mit.edu/",
     "CSRF_COOKIE_DOMAIN": f".{mitlearn_config.get('frontend_domain')}",
+    "DEFAULT_SEARCH_MAX_INCOMPLETENESS_PENALTY": 0,
+    "DEFAULT_SEARCH_STALENESS_PENALTY": 0,
     "EDX_API_ACCESS_TOKEN_URL": "https://api.edx.org/oauth2/v1/access_token",
     "EDX_API_URL": "https://api.edx.org/catalog/v1/catalogs/10/courses",
     "EDX_PROGRAMS_API_URL": "https://discovery.edx.org/api/v1/programs/",
+    "EMBEDDINGS_EXTERNAL_FETCH_USE_WEBDRIVER": True,
+    "EMBEDDING_SCHEDULE_MINUTES": 120,
+    "LITELLM_TOKEN_ENCODING_NAME": "cl100k_base",
     "MICROMASTERS_CATALOG_API_URL": "https://micromasters.mit.edu/api/v0/catalog/",
     "MICROMASTERS_CMS_API_URL": "https://micromasters.mit.edu/api/v0/wagtail/",
     "MITOL_ADMIN_EMAIL": "cuddle-bunnies@mit.edu",
     "MITOL_AXIOS_BASE_PATH": f"https://{mitlearn_config.get('frontend_domain')}",
     "MITOL_DB_CONN_MAX_AGE": 0,
     "MITOL_DB_DISABLE_SSL": "True",
-    "MITOL_SECURE_SSL_REDIRECT": "False",
-    "QDRANT_ENABLE_INDEXING_PLUGIN_HOOKS": True,
     "MITOL_DEFAULT_SITE_KEY": "micromasters",
     "MITOL_EMAIL_PORT": 587,
     "MITOL_EMAIL_TLS": "True",
-    "EMBEDDINGS_EXTERNAL_FETCH_USE_WEBDRIVER": True,
     "MITOL_ENVIRONMENT": env_name,
     "MITOL_FROM_EMAIL": "MIT Learn <mitopen-support@mit.edu>",
     "MITOL_FRONTPAGE_DIGEST_MAX_POSTS": 10,
-    "MITOL_USE_S3": "True",
     "MITOL_NOTIFICATION_EMAIL_BACKEND": "anymail.backends.mailgun.EmailBackend",
+    "MITOL_SECURE_SSL_REDIRECT": "False",
+    "MITOL_USE_S3": "True",
     "MITPE_BASE_URL": "https://professional.mit.edu/",
     "MITX_ONLINE_LEARNING_COURSE_BUCKET_NAME": "mitx-etl-mitxonline-production",
     "NEW_RELIC_LOG": "stdout",
     "NODE_MODULES_CACHE": "False",
     "OCW_BASE_URL": "https://ocw.mit.edu/",
-    "LITELLM_TOKEN_ENCODING_NAME": "cl100k_base",
     "OCW_CONTENT_BUCKET_NAME": "ocw-content-storage",
-    "OCW_UPLOAD_IMAGE_ONLY": "True",
     "OCW_LIVE_BUCKET": "ocw-content-live-production",
+    "OCW_UPLOAD_IMAGE_ONLY": "True",
     "OLL_ALT_URL": "https://openlearninglibrary.mit.edu/courses",
     "OLL_API_ACCESS_TOKEN_URL": "https://openlearninglibrary.mit.edu/oauth2/access_token/",
     "OLL_API_URL": "https://discovery.openlearninglibrary.mit.edu/api/v1/catalogs/1/courses/",
@@ -1013,16 +1017,17 @@ env_vars = {
     "OLL_LEARNING_COURSE_BUCKET_PREFIX": "open-learning-library/courses",
     "OPENSEARCH_DEFAULT_TIMEOUT": 30,
     "OPENSEARCH_INDEXING_CHUNK_SIZE": 75,
+    "POSTHOG_ENABLED": "True",
+    "POSTHOG_PROJECT_ID": 63497,
+    "POSTHOG_TIMEOUT_MS": 1000,
+    "PROLEARN_CATALOG_API_URL": "https://prolearn.mit.edu/graphql",
     "QDRANT_COLLECTION_NAME": f"mitlearn-{stack_info.env_suffix}",
     "QDRANT_DENSE_MODEL": "text-embedding-3-large",
+    "QDRANT_ENABLE_INDEXING_PLUGIN_HOOKS": True,
     "QDRANT_ENCODER": "vector_search.encoders.litellm.LiteLLMEncoder",
-    "CONTENT_FILE_EMBEDDING_CHUNK_OVERLAP": 51,
-    "CONTENT_FILE_EMBEDDING_CHUNK_SIZE": 512,
-    "EMBEDDING_SCHEDULE_MINUTES": 120,
-    "PROLEARN_CATALOG_API_URL": "https://prolearn.mit.edu/graphql",
-    "SEE_API_URL": "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/api/",
-    "SEE_API_ACCESS_TOKEN_URL": "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/oauth/token",
     "SECURE_CROSS_ORIGIN_OPENER_POLICY": "None",
+    "SEE_API_ACCESS_TOKEN_URL": "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/oauth/token",
+    "SEE_API_URL": "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/api/",
     "SEE_BASE_URL": "https://executive.mit.edu/",
     "SOCIAL_AUTH_OL_OIDC_KEY": "ol-mitlearn-client",
     "USE_X_FORWARDED_HOST": "True",
@@ -1030,12 +1035,9 @@ env_vars = {
     "XPRO_CATALOG_API_URL": "https://xpro.mit.edu/api/programs/",
     "XPRO_COURSES_API_URL": "https://xpro.mit.edu/api/courses/",
     "XPRO_LEARNING_COURSE_BUCKET_NAME": "mitx-etl-xpro-production-mitxpro-production",
+    "YOUTUBE_CONFIG_URL": "https://raw.githubusercontent.com/mitodl/open-video-data/mitopen/youtube/channels.yaml",
     "YOUTUBE_FETCH_SCHEDULE_SECONDS": 14400,
     "YOUTUBE_FETCH_TRANSCRIPT_SCHEDULE_SECONDS": 21600,
-    "YOUTUBE_CONFIG_URL": "https://raw.githubusercontent.com/mitodl/open-video-data/mitopen/youtube/channels.yaml",
-    "POSTHOG_ENABLED": "True",
-    "POSTHOG_TIMEOUT_MS": 1000,
-    "POSTHOG_PROJECT_ID": 63497,
 }
 
 # Values that require interpolation or other special considerations
@@ -1324,7 +1326,6 @@ redis_cluster_security_group = ec2.SecurityGroup(
         ec2.SecurityGroupIngressArgs(
             security_groups=[
                 mitlearn_app_security_group.id,
-                operations_vpc["security_groups"]["celery_monitoring"],
                 cluster_substructure_stack.require_output(
                     "cluster_keda_security_group_id"
                 ),
@@ -1333,6 +1334,13 @@ redis_cluster_security_group = ec2.SecurityGroup(
             from_port=DEFAULT_REDIS_PORT,
             to_port=DEFAULT_REDIS_PORT,
             description="Allow application pods to talk to Redis",
+        ),
+        ec2.SecurityGroupIngressArgs(
+            cidr_blocks=operations_vpc["k8s_pod_subnet_cidrs"],
+            protocol="tcp",
+            from_port=DEFAULT_REDIS_PORT,
+            to_port=DEFAULT_REDIS_PORT,
+            description="Allow Operations VPC celery monitoring pods to talk to Redis",
         ),
     ],
     vpc_id=apps_vpc["id"],
